+++ src/main/kotlin/lotto/Application.kt
@@ -2,7 +2,7 @@ package lotto
 
 import lotto.domain.Lotto
 import lotto.domain.LottoResult
-import lotto.service.LottoStore
+import lotto.domain.LottoStore
 import lotto.util.getIntegersAfterSplit
 import lotto.view.InputView
 import lotto.view.OutputView
