+++ src/test/kotlin/lotto/domain/LottoStoreTest.kt
@@ -1,6 +1,5 @@
-package lotto.service
+package lotto.domain
 
-import lotto.domain.Lotto
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
